// apps/web/src/api/generated/models/recentScriptDto.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Script Factory API
 * API for UGC Script Factory
 * OpenAPI spec version: 1.0
 */

export interface RecentScriptDto {
  id: string;
  hook: string;
  angle: string;
  score?: number;
  createdAt: string;
  batchId: string;
  projectId: string;
  projectName: string;
}